Buttons to open input and output directories, interface improvements.

+++ Interface/RenderWindow.py
@@ -24,33 +24,33 @@ class RenderWindow(Window):
         super().window_parameters('Render', 'lightgrey', 500, 250)
 
         # Title Label:
-        lbl_title = QLabel('Render', self)
-        lbl_title.setStyleSheet("font-weight: bold; font-size: 18px;")
-        lbl_title.setGeometry(200, 20, 100, 70)
+        # lbl_title = QLabel('Render', self)
+        # lbl_title.setStyleSheet("font-weight: bold; font-size: 18px;")
+        # lbl_title.setGeometry(200, 20, 100, 70)
 
         # Progress Bar:
         self.pb_progress = QProgressBar(self)
-        self.pb_progress.setGeometry(100, 100, 300, 50)
+        self.pb_progress.setGeometry(100, 20, 300, 50)
         self.pb_progress.setRange(0, 100)
         self.pb_progress.setStyleSheet("border: 2px solid #2196F3; border-radius: 5px; background-color: #E0E0E0;")
 
         # Buttons:
-        self.btn_go = self.button_config('▶', 'green', 'Arial', 30, tooltip_text='Go')
-        self.btn_go.setGeometry(100, 160, 30, 30)
-        self.btn_go.clicked.connect(self.start_rendering)
+        self.btn_start = self.button_config('▶', 'green', 'Arial', 30, tooltip_text='Start')
+        self.btn_start.setGeometry(105, 75, 35, 35)
+        self.btn_start.clicked.connect(self.start_rendering)
 
-        btn_stop = self.button_config('■', 'lightcoral', 'Arial', 30, tooltip_text='Stop')
-        btn_stop.setGeometry(155, 160, 30, 30)
+        btn_stop = self.button_config('⬛', 'lightcoral', 'Arial', 15, tooltip_text='Stop', style='padding-top: 0px; padding-bottom: 2px; padding-left: 0.4px;')
+        btn_stop.setGeometry(145, 75, 35, 35)
         btn_stop.clicked.connect(self.stop_rendering)
 
-        btn_exit = self.button_config('Back', 'lightblue', 'Arial', 12, tooltip_text='Exit')
-        btn_exit.setGeometry(210, 160, 50, 30)
-        btn_exit.clicked.connect(self.close)
-        btn_exit.clicked.connect(self.stop_rendering)
-        btn_exit.clicked.connect(self.controller.execute_main_window)
+        btn_back = self.button_config('Back', 'lightblue', 'Arial', 12, tooltip_text='Back to the main window')
+        btn_back.setGeometry(430, 205, 60, 35)
+        btn_back.clicked.connect(self.close)
+        btn_back.clicked.connect(self.stop_rendering)
+        btn_back.clicked.connect(self.controller.execute_main_window)
 
         self.btn_toggle_delete = self.button_config('Delete: OFF', 'lightcoral', 'Arial', 12, tooltip_text='Delete original video')
-        self.btn_toggle_delete.setGeometry(270, 160, 120, 30)
+        self.btn_toggle_delete.setGeometry(265, 75, 130, 35)
         self.btn_toggle_delete.setCheckable(True)
         self.btn_toggle_delete.clicked.connect(self.change_toggle_delete)
 
@@ -68,10 +68,10 @@ class RenderWindow(Window):
 
     def start_rendering(self):
         self.btn_toggle_delete.setEnabled(False)
-        self.btn_go.setEnabled(False)
+        self.btn_start.setEnabled(False)
 
         # Instances of the threads
-        self.render_thread = RenderThread(self.file_manager, self.slicer, self.pb_progress, self.btn_toggle_delete, self.btn_go)
+        self.render_thread = RenderThread(self.file_manager, self.slicer, self.pb_progress, self.btn_toggle_delete, self.btn_start)
         self.progress_thread = ProgressThread(self.pb_progress)
 
         # Connect the signals
@@ -89,13 +89,14 @@ class RenderWindow(Window):
 
 
     def stop_rendering(self):
-        self.render_thread.terminate()
-        self.progress_thread.terminate()
+        if self.render_thread and self.progress_thread is not None:
+            self.render_thread.terminate()
+            self.progress_thread.terminate()
 
-        self.btn_toggle_delete.setEnabled(True)
-        self.btn_go.setEnabled(True)
+            self.btn_toggle_delete.setEnabled(True)
+            self.btn_start.setEnabled(True)
 
-        self.update_progress(0)
+            self.update_progress(0)
 
     def update_progress(self, value):
         self.pb_progress.setValue(value)
